backend/pyproject.toml
[project]
name = "rag-backend"
version = "1.0.0"
description = "RAG (Retrieval-Augmented Generation) Backend API"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Nitin Singh", email = "nitin27may@gmail.com" }
]

dependencies = [
    # Core API
    "fastapi>=0.115.5",
    "uvicorn[standard]>=0.32.1",
    "python-multipart>=0.0.12",
    "pydantic>=2.10.3",
    "pydantic-settings>=2.6.1",
    "httpx>=0.28.1",
    
    # Environment
    "python-dotenv>=1.0.1",
    
    # Database with pgvector
    "sqlalchemy>=2.0.36",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.6",
    "psycopg[binary]>=3.2.3",
    "psycopg2-binary>=2.9.10",
    
    # LangChain ecosystem
    "langchain>=0.3.14",
    "langchain-core>=0.3.29",
    "langchain-community>=0.3.13",
    "langchain-openai>=0.2.14",
    "langchain-text-splitters>=0.3.4",
    "langchain-postgres>=0.0.16",
    
    # OpenAI
    "openai>=1.58.1",
    "tiktoken>=0.8.0",
    
    # Document processing
    "pypdf>=5.1.0",
    "PyMuPDF>=1.25.1",
    "python-docx>=1.1.2",
    "openpyxl>=3.1.5",
    "markdown>=3.7",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.3.0",
    "pytesseract>=0.3.13",
    "Pillow>=11.0.0",
    "html2text>=2024.2.26",
    "pandas>=2.2.3",
    "requests>=2.32.3",
    
    # Embeddings
    "sentence-transformers>=3.3.1",
    "transformers>=4.47.1",
    "torch>=2.5.1",
    
    # Object storage
    "minio>=7.2.10",
    
    # Utilities
    "tenacity>=9.0.0",
    "numpy>=1.26.4",
    "aiofiles>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "black>=24.0.0",
    "isort>=5.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
